+++ torchtext/prototype/models/t5/__init__.py
@@ -4,6 +4,7 @@ from .bundler import (
     T5Bundle,
 )
 from .model import T5Conf, T5Model
+from .t5_transform import T5Transform
 
 __all__ = [
     "T5Conf",
@@ -11,4 +12,5 @@ __all__ = [
     "T5Bundle",
     "T5_BASE_ENCODER",
     "T5_BASE",
+    "T5Transform",
 ]
